feat(cli.py): add tool to filter and reformat bam for downstream analyses eg methylation
To make the bwa-produced BAM suitable for WhatsHap we add a suffix to the rname to ensure that each alignment is unique. We need to strip this out for tools that need to match the alignment up with the fast5 record (eg f5c/nanopolish). We also need to filter the raw bam to remove alignments that fail the pore-c filters.

(cherry picked from commit 74b70e9f242d04b462d632eb94f4b723709049a6)