feat(pore_c/analyses/variants.py): add a tool to link variants in a vcf using pore_c reads
A useful datatype to have access to is the count of pore-c reads that link a pair of variant in cis or trans. This tool creates a table with these counts, it is roughly equivalent to hapcut2's extractHairs. This change also required a more recent version of dask, so I took the opportunity to clean up the dask parameters in the cli.